requirements.txt
requests
pymupdf
pandas
streamlit
python-dotenv
google-generativeai
feedparser